+++ lib/src/collect.dart
@@ -217,30 +217,9 @@ Future<List<Map<String, dynamic>>> _getCoverageJson(VmService service,
   // Output JSON
   final coverage = <Map<String, dynamic>>[];
   hitMaps.forEach((uri, hitMap) {
-    coverage.add(_toScriptCoverageJson(uri, hitMap));
+    coverage.add(toScriptCoverageJson(uri, hitMap));
   });
   return coverage;
-}
-
-/// Returns a JSON hit map backward-compatible with pre-1.16.0 SDKs.
-Map<String, dynamic> _toScriptCoverageJson(
-    Uri scriptUri, Map<int, int> hitMap) {
-  final json = <String, dynamic>{};
-  final hits = <int>[];
-  hitMap.forEach((line, hitCount) {
-    hits.add(line);
-    hits.add(hitCount);
-  });
-  json['source'] = '$scriptUri';
-  json['script'] = {
-    'type': '@Script',
-    'fixedId': true,
-    'id': 'libraries/1/scripts/${Uri.encodeComponent(scriptUri.toString())}',
-    'uri': '$scriptUri',
-    '_kind': 'library',
-  };
-  json['hits'] = hits;
-  return json;
 }
 
 class StdoutLog extends Log {